requirements.txt
discord.py==2.0.1
python-dotenv==0.19.2
Flask==2.1.0
Werkzeug==2.0.2
requests